### 2020-07-09 #2458 @whedon
[PRE REVIEW]: pirecorder: Controlled and automated image and video recording with the raspberry pi
**Submitting author:** @JolleJolles (<a href="http://orcid.org/0000-0001-9905-2633">Jolle Wolter Jolles</a>)
**Repository:** <a href="http://github.com/jollejolles/pirecorder" target ="_blank">http://github.com/jollejolles/pirecorder</a>
**Version:** v3.1.0
