### 2020-07-09 #2458 @kyleniemeyer
Hi @JolleJolles, thanks for your submission to JOSS. As we are getting started, could you please revise your paper to include an explicit Statement of Need section, as described in [our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)?
