+++ paper/paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 ---
 
 # Summary
-A fundamental component of empirical research is the acquisition of accurate, consistent, and often significant amounts of data. The raspberry pi is small, single-board computer that facilitates controlled and automated data collection by bringing together open-source principles with sensor and controller interfaces, and highly customisable programming capabilities. In recent years, the raspberry pi has also been increasingly taken up by the scientific community [@Fletcher2019], accelerated by its low cost, ease of use, and high flexibility. One of its foremost uses is the collection of high quality image data, from sub-micron resolution microscopy [@Aidukas2019], and deep sea video recordings [@Phillips2019], to motion-triggered camera trapping [@Nazir2017; @Prinz2016], high-throughput behavioural assessments [@Geissmann2017; @Jolles2019; @Todd2017], long-term home cage monitoring [@Singh2019], and the automated tracking of animal groups [@Alarcon-Nieto2018; @Jolles2018].
+A fundamental component of empirical research is the acquisition of accurate, consistent, and often significant amounts of data. Specifically, researchers often require large numbers of controlled and often parallel image and video recordings. For this the raspberry pi, a small, single-board computer that brings together open-source principles with sensor and controller interfaces, and highly customisable programming capabilities, provides a great, low cost solution. Indeed, in recent years, the raspberry pi has been increasingly taken up by the scientific community [@Fletcher2019] and used in a wide range of projects that required the collection of high quality image data, from sub-micron resolution microscopy [@Aidukas2019], and deep sea video recordings [@Phillips2019], to motion-triggered camera trapping [@Nazir2017; @Prinz2016], high-throughput behavioural assessments [@Geissmann2017; @Jolles2019; @Todd2017], long-term home cage monitoring [@Singh2019], and the automated tracking of animal groups [@Alarcon-Nieto2018; @Jolles2018].
 
-Researchers can use the native software that comes with the raspberry pi to take still photographs and videos directly from the command line (`raspistill` and `raspivid`) or control the camera module using Python with the extensive `picamera` Python package [@Jones2017] and write their own recording scripts. Also specific solutions exist for the monitoring of camera signals and the detection of motion ([Motion](https://motion-project.github.io)), home cage observations via a web-based interface [@Singh2019], and behavioural profiling with the raspberry pi [@Geissmann2017]. However, a flexible, complete solution for controlled and automated image and video recording with the raspberry pi was still lacking.
+So far, researchers have often relied on writing their own recordings scripts to take still photographs and videos from the command line (using `raspistill` and `raspivid`), control the camera module with `picamera` in Python [@Jones2017], or trigger recordings by motion-detection  ([Motion](https://motion-project.github.io)). Also some specific solutions exist, such as a web-based interface to run recordings [@Singh2019] and advanced software that converts the raspberry pi in a dedicated behavioural profiling machine [@Geissmann2017]. What is still missing is a complete solution that helps researchers set up their raspberry pi and camera and configure camera and recording parameters to enable them to run large numbers of (parallel and automated) image and video recordings in a simple way. Here I present `pirecorder` to overcome this need.
 
[…]
