### 2020-07-13 #2458 @JolleJolles
> Hi @JolleJolles, thanks for your submission to JOSS. As we are getting started, could you please revise your paper to include an explicit Statement of Need section, as described in [our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)?

Dear @kyleniemeyer, thanks for your help. I have now updated ([commit](https://github.com/JolleJolles/pirecorder/commit/c2f23e4376c21bcafd6588e83650b258673644ab)) the paper to include a more explicit statement of need .
