### 2020-07-15 #2458 @marcosvital
Hi, @JolleJolles, I will be the editor of your submission.

I'll generate a second version of your manuscript proof, so you can start by checking if everything seems to be ok with it, as it will be the version used to start revision if nothing else changes. If you make any changes, you can use the same whedon command I will use below to create a new proof (or just let me know and I'll do it).

At the same time, we can start looking for reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet).
You can also check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.
