### 2020-07-20 #2458 @marcosvital
Dear @lucask07, would you be willing to review this submission for JOSS?

This is still a pre-review, once there are enough reviewers, we will move on to a new issue where the review will take place.

If you are not able to review, please let us know if you have recommendations for other reviewers.
