### 2020-07-20 #2458 @lucask07
Hi, @marcosvital I am willing to review but do not have a RaspberryPi, camera combo. I'll let you decide if the ability to run the code completely with the appropriate hardware is required of the reviewers.
