### 2020-07-29 #2458 @marcosvital
@lucask07, I think it will be ok with you not having the hardware if we got other reviewers to do that, so I would love if you could review this submission. Let me know if you are still available for that - we are likely to start next week.
