### 2020-07-29 #2458 @JolleJolles
Great. Looking forward.

> On 29 Jul 2020, at 13:45, Marcos Vital (LEQ-UFAL) <notifications@github.com> wrote:
>
> ﻿
> Hi, @JolleJolles, a quick update: Jan Heuschele agreed to be one of the reviewers. He is out of office right now, but it'll return in a week. My initial intent is to wait a little more to start review so we can have him, are you ok with that?
>
