### 2020-08-20 #2584 @JolleJolles
Thanks everyone for your help. Looking forward.

> On 20 Aug 2020, at 12:46, Marcos Vital (LEQ-UFAL) <notifications@github.com> wrote:
>
>
> Dear @JolleJolles <https://github.com/JolleJolles>: your manuscript will be reviewed in this issue, and you can reply any comments and suggestions that the reviewers might address right here.
>
