### 2020-08-21 cdde696
improved imseq timing
