+++ pirecorder/pirecorder.py
@@ -588,6 +588,7 @@ class PiRecorder:
                 if i < self.config.img.imgnr-1:
                     #timepassed = (datetime.now() - timept2).total_seconds()
                     #delay = max(0, self.config.img.imgwait - timepassed)
+                    print(self.filename)
                     delay = self.config.img.imgwait-(timept2.microsecond/1000000)
                     lineprint("Captured "+img+", sleeping "+str(round(delay,2))+"s..")
                     sleep(delay)
