+++ pirecorder/pirecorder.py
@@ -586,10 +586,11 @@ class PiRecorder:
                                     format="jpeg", resize = self.resize,
                                     quality = self.config.img.imgquality)):
                 if i < self.config.img.imgnr-1:
-                    #timepassed = (datetime.now() - timept2).total_seconds()
-                    #delay = max(0, self.config.img.imgwait - timepassed)
-                    timeimgtaken = int(img.split("_")[-1:][0][6:-4])/1000000
-                    delay = self.config.img.imgwait-timeimgtaken
+                    timepassed = (datetime.now() - timept2).total_seconds()
+                    delay = max(0, self.config.img.imgwait - timepassed)
+                    #timeimgtaken = int(img.split("_")[-1:][0][6:-4])/1000000
+                    #delay = self.config.img.imgwait-timeimgtaken
+                    datetime.now()
                     lineprint("Captured "+img+", sleeping "+str(round(delay,2))+"s..")
                     sleep(delay)
                     timept2 = datetime.now()
