+++ pirecorder/pirecorder.py
@@ -593,7 +593,6 @@ class PiRecorder:
                     lineprint("Captured "+img+", sleeping "+str(round(delay,2))+"s..")
                     sleep(delay)
                     timept2 = datetime.now()
-                    print(timept2)
                 else:
                     lineprint("Captured "+img)
                     break
