### 2020-08-26 #2584 @DerJH
Hi,
I am trying to install pirecorder at the moment, but both using _sudo pip install pirecorder_ and the GitHub version i am running into the following error code (I attached the whole commandline dialog as a text file). It seems to have something to do with "h5py". I then tried to install with "sudo pip install h5py", which also failed.

Command "/usr/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-eCsuDf/h5py/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" install --record /tmp/pip-record-Vphw1e/install-record.txt --single-version-externally-managed --compile" failed with error code 1 in /tmp/pip-install-eCsuDf/h5py/

Is there an obvious workaround or fix for that?
Best, Jan
[full_error.txt](https://github.com/openjournals/joss-reviews/files/5129752/full_error.txt)
