### 2020-08-26 #2584 @JolleJolles
Hi Jan,

That seems to be an error with the h5py package, not with pirecorder. I have never encountered it. I have installed pirecorder on fresh raspbian systems on rpi2,3, and 4 and then do not get this error. I am not sure how to solve the h5 install error, perhaps try updating your raspberry pi?

h5py is not actually used by pirecorer but only by 1 function in my [pythutils package](https://github.com/JolleJolles/pythutils) that is also not used by pirecorder. I updated pythutils to not require the h5py package anymore, so hopefully now when installing pythutils and pirecorder you can get past this stage. You have to use the github version (1.3.6). I will update it on pythutils after the package and paper are passed the review stage.

I hope this helps, do let me know if something else related to this point has come up.

Thanks,
Jolle
