### 2020-08-26 #2584 @DerJH
Hi Jolle,
ok after upgrading raspbian and using apt-get python-h5py and apt-get python3-h5py the installation worked!
I'll try to test all the functions in your program in the next 2 days.
Greetings,
 Jan
