+++ pirecorder/__version__.py
@@ -15,4 +15,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-__version__ = "3.1.0"
+__version__ = "3.1.x"
