### 2020-08-27 #2584 @lucask07
Hi @JolleJolles,
A quick question before I dig too deep. Is your package targeting a specific Raspberry Pi compatible camera? Have you documented the cameras your software has been tested with? I suspect the compatible cameras are set by the Picamera package. I see these "official" Raspberry Pi cameras that connect via the MIPI port [here](https://www.raspberrypi.org/documentation/hardware/camera/README.md)
