### 2020-08-28 #2584 @JolleJolles
Hi @lucask07. No the package does not target a specific camera and should work with all cameras, official and non-official that connect with the MIPI port as it builds on the picamera package.
Thanks
