### 2020-08-31 #2584 @lucask07
Hi @marcosvital
It's working, I can click on the check-boxes now.
