### 2020-08-31 #2584 @lucask07
@JolleJolles

A few comments as I'm working through the **Documentation** section of the reviewer checklist.

- README page is missing an explanation of how others may "1) Contribute to the software"
- The _Statement of Need_ should be elaborated on in both the README and the github.io page.
- The software does not have automated tests. "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"
