### 2020-08-31 #2584 @lucask07
For the scholarly contribution a check with [CLOC](https://github.com/AlDanial/cloc) ("count lines of code") shows:

- Python: 1093 lines of code, and 402 lines of comment
- Markdown 1150 lines of code
- Jupyter Notebook: 52 lines of code, 973 lines of comment

So this work is above the LOC threshold. Submissions with LOC less than 1000 are "flagged" per the JOSS [scholarly effort guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).
