### 2020-09-01 #2584 @JolleJolles
> For the scholarly contribution a check with [CLOC](https://github.com/AlDanial/cloc) ("count lines of code") shows:
>
> * Python: 1093 lines of code, and 402 lines of comment
> * Markdown 1150 lines of code
> * Jupyter Notebook: 52 lines of code, 973 lines of comment
>
> So this work is above the LOC threshold. Submissions with LOC less than 1000 are "flagged" per the JOSS [scholarly effort guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

Thanks. I have actually very limited comments throughout my code. The only commented code would be the function documentation. Could you guide me on what to here? Thank you.
