### 2020-09-01 #2584 @DerJH
> > @JolleJolles
> > A few comments as I'm working through the **Documentation** section of the reviewer checklist.
> >
> > * README page is missing an explanation of how others may "1) Contribute to the software"
> > * The _Statement of Need_ should be elaborated on in both the README and the github.io page.
> > * The software does not have automated tests. "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"
>
> Thanks @lucask07. I will address your comments 1 and 2 immideately.
>
> For Comment 3, I read in the author guidelines that automated tests are not a requirement and that "Documented manual steps that can be followed to objectively check the expected functionality of the software" would also be okay. I have detailed tutorials to run through the whole software (jollejolles.github.io/pirecorder), which I thought, perhaps wrongly so, that that would conform to that requirement. Would that indeed be enough if I would add a clear statement that the software can be tested by following the online manual? Or alternatively if I add python files that can be run to show the software is working? I would appreciate your feedback on this so I can address the comment appropriately.

Hi, I think a python file which executes several functions would be a good idea, but as I am also new to the software testing business (and github reviewing)  not sure if that is the way these things go about.
