### 2020-09-01 #2584 @DerJH
I have some more questions/comments

1. Is there a way to save the ROI directly into the config file? So 1. opening stream, 2. making rectangle, 3. press "s", 4. save to the current .conf file?

2. In the convert function I might have discovered a tiny bug.

from pirecorder import Convert
Convert(indir = "/home/pi/Videos", outdir = "/home/pi/Videos/converted", resizeval = 0.5)
Convert(indir = "/home/pi/Videos", outdir = "/home/pi/Videos/converted", resizeval = 0.5, withframe = True)

when i do the "withframe = True" then the result becomes a scrambled video mostly grey and random pix, but it works without.

3. The following example on the website has a small error. Twice "imgwait":
rec.settings(imgwait = 1, imgnr = 10, imgtime = 15, imgwait = 60, imgquality = 90 \
             imgdims = (3280, 2464))

4.This is more a wish for a future update: On a small screen the pirecorder.Camconfig() slider window only allows access until the first Red channel slider. Maybe this could be made more responsive, or all the UI elements  a bit smaller.

5. all the above is tested using python 2.7. I somehow didn't manage to get it to work under python 3, but given my general linux skills this is probably more a problem of me failing to manage to install it properly. But maybe you can help me to figure out what could be wrong here.
pi@raspberrypi:~ $ python3
Python 3.7.3 (default, Jul 25 2020, 13:03:44)
[GCC 8.3.0] on linux
>>> import cv2
>>> import pirecorder
>>> rec = pirecorder.PiRecorder(configfile = "irsettings.conf")
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'pirecorder' has no attribute 'PiRecorder'
>>> exit()

thanks
