### 2020-09-09 #2584 @lucask07
Hi @JolleJolles,
I'm going through the paper and have a few concerns/nitpicks that I'd suggest you clarify.

- The paper references "get the optimal camera shutter speed". This is a rather complex calculation that attempts to limit saturation while simultaneously having the darkest pixels collect the most possible light. I can't find any auto exposure calculations in your code. Can you clarify how the optimal shutter speed is calculated/retrieved?
- The paper reference “Calibrating the camera.” This is not appropriate usage of the term **_calibrate_** for a camera. Camera calibration may involve determining the focal length and image center or estimating the lens distortion. Other calibration efforts may center around aspects of sensor performance such as measuring the color correction matrix or locating hot pixels. This does not appear to be what _pirecorder_ is doing. I think your package enables real-time streaming so that a user is given feedback while positioning a camera? Can you change the paper wording to clarify what _pirecorder_ helps the user accomplish?
- _Paper grammar_: “Multiple configuration files can be used that can be edited directly, such as to set the video duration, quality, or number of images, or dynamically via an interactive video stream, such as to draw the region of interest, set the desired white-balance, or get the optimal camera shutter speed.”
The “or”s in this list are confusing. Should these actually be “and”s? Consider splitting this sentence into 2 or 3 sentences for clarity.
- The [Jupyter notebook](https://github.com/JolleJolles/pirecorder/blob/9b54d9cd77996e5387072d5be3313f804782d564/notebooks/running-PiRecorder.ipynb) references the three possible cameras to use:
1. v1.5 camera model (2592x1944)
2. v2 camera (3280x2464)
3. HQ camera (4056x3040)
I think these 3 cameras should be more clearly called out in the README.
- The code _camconfig.py_ allows for a limited range of ISO.
```
isos = [100,200,320,400,500,640,800]
```
Is this consistent with the capabilities of the 3 camera modules? Is it possible to probe the camera module for the available ISO settings?
- The code _camconfig.py_ appears to limit resolution to 1640x1232:
```
if (res[0]*res[1])>=(1640*1232):
        res = maxrect(dims, maxdims = (1640,1232), decimals = 0)
cam.resolution = picamconv(res)
```
Is my interpretation correct? Is this limitation required?

Thanks for all your work.
