### 2020-09-10 #2584 @JolleJolles
Thanks for all your comments. I will address them all together from next Monday (14th), making sure to add each point as an issue to the repository.
Thanks
