### 2020-09-14 JolleJolles/pirecorder#22 @JolleJolles
Clarify camera compatibility
Clarify pirecorder should work with all cameras that connect to the MIPI port and that models can be set in the config file to determine max resolution.

In response to @lucask07 in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-682057040_

> Hi @JolleJolles,
> A quick question before I dig too deep. Is your package targeting a specific Raspberry Pi compatible camera? Have you documented the cameras your software has been tested with? I suspect the compatible cameras are set by the Picamera package. I see these "official" Raspberry Pi cameras that connect via the MIPI port [here](https://www.raspberrypi.org/documentation/hardware/camera/README.md)
