### 2020-09-14 JolleJolles/pirecorder#23 @JolleJolles
Clarify statement of need
Clarify the statement of need in JOSS paper, README, and on the documentation website.

In response to @lucask07 in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684089120_
