### 2020-09-14 JolleJolles/pirecorder#24 @JolleJolles
Add potential contribution statement
Add an explanation to the README how others may contribute to the software.

In response to @lucask07 in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684089120_
