### 2020-09-14 JolleJolles/pirecorder#25 @JolleJolles
Add software tests
Include Python script for testing all functionality of pirecorder and add corresponding statements how testing can be done in the README and the documentation pages.

In response to @lucask07 in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684089120_
