### 2020-09-14 JolleJolles/pirecorder#26 @JolleJolles
Save ROI directly using stream from command line
Add option to run stream function from command line with configfile such that roi file can be stored directly

In response to @DerJH in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684568533_
