### 2020-09-14 JolleJolles/pirecorder#27 @JolleJolles
Clarify issue with opencv video conversion
In response to issue highlighted by @DerJH in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684568533_
