### 2020-09-14 JolleJolles/pirecorder#28 @JolleJolles
Possibility for improved slider window sizing
In response to @DerJH in https://github.com/openjournals/joss-reviews/issues/2584#issuecomment-684568533_
