### 2020-09-14 41117c5
Statement about minimum resolution. Closes #28
