### 2020-09-14 JolleJolles/pirecorder#28 @JolleJolles
@DerJH In response to your request, the user interface is built using OpenCV which does not have any (obvious) option to resize the elements. Furthermore, setting the raspberry pi to the smallest (selectable) resolution of 640x480 pixels all the scrollbars of the UI are visible so I cannot reproduce your issue.

It is recommended to have a minimum resolution of 800x600 pixels, and 1280x1024 is preferable. This clarification is now added to the documentation.
