### 2020-09-14 #2584 @JolleJolles
> 1. The following example on the website has a small error. Twice "imgwait":
>    rec.settings(imgwait = 1, imgnr = 10, imgtime = 15, imgwait = 60, imgquality = 90
>    imgdims = (3280, 2464))

Thanks, fixed the typo with commit [c2f23e4](https://github.com/JolleJolles/pirecorder/commit/c2f23e4376c21bcafd6588e83650b258673644ab)
