### 2020-09-14 e11e246
Contribution statement added. Closes #24
