### 2020-09-14 edae82f
Statement of need added. Closes #23
