### 2020-09-14 #2584 @JolleJolles
> A few comments as I'm working through the **Documentation** section of the reviewer checklist.
>
> * README page is missing an explanation of how others may "1) Contribute to the software"
> * The _Statement of Need_ should be elaborated on in both the README and the github.io page.
> * The software does not have automated tests. "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Following the above point, I improved the statement of need in the paper and added the statement to the README and the web documentation (issue #23, commit [edae82f](https://github.com/JolleJolles/pirecorder/commit/edae82f5d6e2836dec39e55602981e8edf5b4e53)).
