### 2020-09-14 #2584 @JolleJolles
> * The code _camconfig.py_ allows for a limited range of ISO.
>
> ```
> isos = [100,200,320,400,500,640,800]
> ```
>
> Is this consistent with the capabilities of the 3 camera modules? Is it possible to probe the camera module for the available ISO settings?

In reply to this point, pirecorder builts on the picamera package to control the native camera settings, which are indeed limited to the above range, see the [pirecorder documentation](https://picamera.readthedocs.io/en/release-1.10/api_camera.html#picamera.camera.PiCamera.iso). pirecorder uses a default iso setting of 200.
