### 2020-09-15 cc6e125
Store roi directly from terminal stream
