+++ pirecorder/stream.py
@@ -232,7 +232,7 @@ def strm():
              description=Stream.__doc__,
              formatter_class=argparse.RawDescriptionHelpFormatter)
 
-    parser.add_argument("-c", "--cameratype", default="v2", metavar="")
+    parser.add_argument("-t", "--cameratype", default="v2", metavar="")
     parser.add_argument("-f", "--framerate", default=8, type=int, metavar="")
     parser.add_argument("-v", "--vidsize", default=0.2, type=float, metavar="")
     parser.add_argument("-r", "--rotation", default=0, type=int, metavar="")
