### 2020-09-15 #2584 @JolleJolles
> All the above is tested using python 2.7. I somehow didn't manage to get it to work under python 3, but given my general linux skills this is probably more a problem of me failing to manage to install it properly. But maybe you can help me to figure out what could be wrong here.
>    pi@raspberrypi:~ $ python3
>    Python 3.7.3 (default, Jul 25 2020, 13:03:44)
>    [GCC 8.3.0] on linux
>    Type "help", "copyright", "credits" or "license" for more information.
>
> > > > import cv2
> > > > import pirecorder
> > > > rec = pirecorder.PiRecorder(configfile = "irsettings.conf")
> > > > Traceback (most recent call last):
> > > > File "", line 1, in
> > > > AttributeError: module 'pirecorder' has no attribute 'PiRecorder'
> > > > exit()

@DerJH in reply to your above question, I cannot reproduce your issue. Doing a fresh install of pirecorder on python3 I do not get this error. Could you show the output of `pirecorder.__version__` and `cv2.__version` so I can help further? I also recommend to use a virtual environment (as I detail in the online documentation)
