### 2020-09-15 JolleJolles/pirecorder#26 @JolleJolles
Integrated saving directly from stream by possibility of adding a configuration file when running stream from the command line (e.g. `stream -c pirecorder.conf`). When the 's' key is pressed after drawing a rectangle, the roi coordinates are stored in the config file. Working after commit [d335de9](https://github.com/JolleJolles/pirecorder/commit/f8528fa642910f69068196be3385b83f20f98c6f).
