### 2020-09-15 #2584 @JolleJolles
> 1. Is there a way to save the ROI directly into the config file? So 1. opening stream, 2. making rectangle, 3. press "s", 4. save to the current .conf file?

Integrated saving directly from stream by the possibility of adding a configuration file when running stream from the command line (e.g. `stream -c pirecorder.conf`). When the 's' key is pressed after drawing a rectangle, the roi coordinates are stored in the config file. Working after commit [d335de9](https://github.com/JolleJolles/pirecorder/commit/f8528fa642910f69068196be3385b83f20f98c6f).
