### 2020-09-15 7145236
Integrated ability to store ROI from stream via command line
