### 2020-09-15 JolleJolles/pirecorder#27 @JolleJolles
I cannot reproduce the issue and it seems it has to do when running an older version of opencv from python2.7, which is no longer supported. I therefore close this issue for now but will reopen if it persists under python3.
