### 2020-09-15 0a75604
fix release of stream for autoconfig
