### 2020-09-16 5d69fb5
fix release of stream
