### 2020-09-16 c3e214a
fix release of stream
