### 2020-09-16 abd99c8
Incorporated tests. Closes #25
