+++ tests/test.py
@@ -1,0 +1,155 @@
+#! /usr/bin/env python
+"""
+Copyright (c) 2020 - 2020 Jolle Jolles <j.w.jolles@gmail.com>
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at:
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+import os
+import time
+import shutil
+from pythutils.fileutils import listfiles
+
+# import the package
+print("TEST: import pirecorder package")
+import pirecorder
+time.sleep(1)
+print("DONE..\n")
+
+# Initialise PiRecorder class with default configfile
+## Sets up the pirecorder directory, creates configfile, stores settings
+print("TEST: initialise PiRecorder class")
+rec = pirecorder.PiRecorder()
+time.sleep(1)
+print("DONE..\n")
+
+# Initialise PiRecorder with custom configuration file
+print("TEST: initialise PiRecorder class with custom configuration file")
+rec = pirecorder.PiRecorder(configfile = "test.conf")
+time.sleep(1)
+print("DONE..\n")
+
+# Change configuration settings and store automatically
+print("TEST: setting all custom configuration settings")
+rec.settings(recdir = "TESTS", label = "TEST", rectype = "img", subdirs = False,
+             rotation = 0, brighttune = 0, gains = (1.0, 2.5), brightness = 45,
+             contrast = 20, saturation = -100, iso = 200, sharpness = 50,
+             compensation = 0, shutterspeed = 10000, imgdims = (3280,2464),
+             viddims = (1640,1232), imgfps = 1, vidfps = 24, imgwait = 1.0,
+             imgnr = 60, imgtime = 60, imgquality = 50, vidduration = 10,
+             viddelay = 0, vidquality = 11)
+time.sleep(1)
+print("DONE..\n")
+
+# Automatically get the configuration Settings
+print("TEST: running auto configuration (shutterspeed and whitebalance)")
+print("Before: shutterspeed = " + str(rec.config.cam.shutterspeed) + "; gains = " + str(rec.config.cus.gains))
+rec.autoconfig()
+time.sleep(1)
+print("DONE..\n")
+
+# Test recording 1: a single image
+print("TEST: recording a single image")
+rec.record()
+print("DONE..\n")
+
+# Test recording 2: a sequence of 10 images, 5s apart
+print("TEST: recording a sequence of 10 images, 5s apart")
+rec.settings(rectype = "imgseq", imgnr = 10, imgtime = 60, imgwait = 5,
+             subdirs = True)
+rec.record()
+print("DONE..\n")
+
+# Test recording 3: a single 10s video
+print("TEST: recording a 10s video")
+rec.settings(rectype = "vid", vidduration = 10, viddelay = 0, subdirs = False)
+rec.record()
+print("DONE..\n")
+
+# Test recording 4: a sequence of videos
+print("TEST: recording a sequence of videos")
+rec.settings(rectype = "vidseq")
+rec.record()
+time.sleep(1)
+print("DONE..\n")
+
+# Run video stream
+print("TEST: run video stream")
+print("Function records mouse clicks and movements and responds to keypresses:")
+print("f-key: Display the stream fullscreen/display the stream in a window")
+print("c-key: Display/hide a diagonal cross across the screen")
[…]
