### 2020-09-16 #2584 @JolleJolles
> A few comments as I'm working through the **Documentation** section of the reviewer checklist.
>
> * README page is missing an explanation of how others may "1) Contribute to the software"
> * The _Statement of Need_ should be elaborated on in both the README and the github.io page.
> * The software does not have automated tests. "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

I have now included a python test file in the package (`tests/test.py`) to run all functionalities, including running pirecorder, setting configuration files, running different types of recordings, scheduling recordings, converting recordings, and testing the stream and camconfig interactive functions. The possibility to test the software is also described in the README and in the online documentation.
