+++ pirecorder/videoin.py
@@ -27,7 +27,7 @@ from pythutils.mediautils import *
 class VideoIn:
 
     def __init__(self, system = "auto", vidsize = 0.2, framerate = 32,
-                 crop = False, rotation = 0, cameratype = None):
+                 crop = False, rotation = 0, maxres = None):
 
         """Opens a video stream from native camera, webcam or rpi camera"""
 
@@ -44,9 +44,11 @@ class VideoIn:
             from picamera.array import PiRGBArray
             from picamera import PiCamera
 
-            if cameratype == "v2":
+            if isinstance(maxres, tuple):
+                self.maxres = maxres
+            elif maxres == "v2":
                 self.maxres = (3264,2464)
-            elif cameratype == "hq":
+            elif maxres == "hq":
                 self.maxres = (4056,3040)
             else:
                 self.maxres = (2592,1952)
