### 2020-09-16 1772b5b
Various improvements
