+++ paper/paper.md
@@ -23,17 +23,17 @@ bibliography: paper.bib
 # Summary
 A fundamental component of empirical research is the acquisition of accurate, consistent, and often significant amounts of data. Specifically, researchers often require large numbers of controlled and often parallel image and video recordings. For this the raspberry pi, a small, single-board computer that brings together open-source principles with sensor and controller interfaces, and highly customisable programming capabilities, provides a great, low cost solution. Indeed, in recent years, the raspberry pi has been increasingly taken up by the scientific community [@Fletcher2019] and used in a wide range of projects that required the collection of high quality image data, from sub-micron resolution microscopy [@Aidukas2019], and deep sea video recordings [@Phillips2019], to motion-triggered camera trapping [@Nazir2017; @Prinz2016], high-throughput behavioural assessments [@Geissmann2017; @Jolles2019; @Todd2017], long-term home cage monitoring [@Singh2019], and the automated tracking of animal groups [@Alarcon-Nieto2018; @Jolles2018].
 
-So far, researchers have often relied on writing their own recordings scripts to take still photographs and videos from the command line (using `raspistill` and `raspivid`), control the camera module with `picamera` in Python [@Jones2017], or trigger recordings by motion-detection  ([Motion](https://motion-project.github.io)). Also some specific solutions exist, such as a web-based interface to run recordings [@Singh2019] and advanced software that converts the raspberry pi in a dedicated behavioural profiling machine [@Geissmann2017]. There is a need for a complete solution that helps researchers, especially those with limited coding skills, to easily set up and configure their raspberry pi and run large numbers of controlled and automated image and video recordings. Here I present `pirecorder` to overcome this need.
+So far, researchers have often relied on writing their own recordings scripts to take still photographs and videos from the command line (using `raspistill` and `raspivid`), control the camera module with `picamera` in Python [@Jones2017], or trigger recordings by motion-detection  ([Motion](https://motion-project.github.io)). Also some specific solutions exist, such as a web-based interface to run recordings [@Singh2019] and advanced software that converts the raspberry pi in a dedicated behavioural profiling machine [@Geissmann2017]. However, there is still a need for a complete solution that helps researchers, especially those with limited coding skills, to easily set up and configure their raspberry pi and run large numbers of controlled and automated image and video recordings. Here I present `pirecorder` to overcome this need.
 
-`pirecorder` is a Python package, built on the picamera [@Jones2017] and OpenCV [@Bradski2000] libraries, that provides a flexible solution for the collection of consistent image and video data. It consists of a number of interconnected modules to facilitate all aspects of media recording, from setting-up, calibrating and configuring the camera, recording of images and videos, time-lapses, and standardised video sequences with automatic file-naming, easy scheduling of future recordings, and converting of recorded media with resize, timestamp, and monitoring options. All modules also work directly from the command line, making it very straightforward to configure, initiate, schedule, and convert recordings. `pirecorder` comes with a dedicated website with detailed documentation and tutorials ([jollejolles.github.io/pirecorder](https://jollejolles.github.io/pirecorder/)) as well as a set of annotated [Jupyter Notebooks](https://github.com/JolleJolles/pirecorder/tree/master/notebooks) to help users integrate the raspberry pi and `pirecorder` in their work.
[…]
