### 2020-09-16 278ebc3
Improvements to paper
