+++ paper/paper.md
@@ -25,11 +25,11 @@ A fundamental component of empirical research is the acquisition of accurate, co
 
 So far, researchers have often relied on writing their own recordings scripts to take still photographs and videos from the command line (using `raspistill` and `raspivid`), control the camera module with `picamera` in Python [@Jones2017], or trigger recordings by motion-detection  ([Motion](https://motion-project.github.io)). Also some specific solutions exist, such as a web-based interface to run recordings [@Singh2019] and advanced software that converts the raspberry pi in a dedicated behavioural profiling machine [@Geissmann2017]. However, there is still a need for a complete solution that helps researchers, especially those with limited coding skills, to easily set up and configure their raspberry pi and run large numbers of controlled and automated image and video recordings. Here I present `pirecorder` to overcome this need.
 
-`pirecorder` is a Python package, built on the picamera [@Jones2017] and OpenCV [@Bradski2000] libraries, that provides a flexible solution for the collection of consistent image and video data. It consists of a number of interconnected modules to facilitate key aspects of media recording: 1) setting-up and configuring the camera, 2) recording images, videos, time-lapses, and standardised video sequences with automatic file-naming, 3) easy scheduling of future recordings, and 4) converting of recorded media with resize, timestamp, and monitoring options. All functionalities are designed to make it very straightforward, even for users with limited coding experience, to configure, initiate, schedule, and convert recordings. In particular, `pirecorder` offers interactive streaming functionality to facilitate users in positioning and focusing the camera, setting the desired white-balance and other image parameters using trackbars, and set the ideal camera shutter speed. Furthermore, `pirecorder` comes with a dedicated documentation website with detailed information and tutorials ([jollejolles.github.io/pirecorder](https://jollejolles.github.io/pirecorder/)) as well as a set of annotated [Jupyter Notebooks](https://github.com/JolleJolles/pirecorder/tree/master/notebooks) to help users integrate the raspberry pi and `pirecorder` in their work.
+`pirecorder` is a Python package, built on the picamera [@Jones2017] and OpenCV [@Bradski2000] libraries, that provides a flexible solution for the collection of consistent image and video data. It consists of a number of interconnected modules to facilitate key aspects of media recording: 1) setting-up and configuring the camera, 2) recording images, videos, time-lapses, and standardised video sequences with automatic file-naming, 3) easy scheduling of future recordings, and 4) converting of recorded media with resize, timestamp, and monitoring options. All functionalities are designed to make it very straightforward, even for users with limited coding experience, to configure, initiate, schedule, and convert recordings. In particular, `pirecorder` offers interactive streaming functionalities to facilitate users in positioning and focusing the camera, selecting the desired white-balance and other image parameters using trackbars, and set the ideal camera shutter speed. Furthermore, `pirecorder` comes with a dedicated documentation website with detailed information and tutorials ([jollejolles.github.io/pirecorder](https://jollejolles.github.io/pirecorder/)) as well as a set of annotated [Jupyter Notebooks](https://github.com/JolleJolles/pirecorder/tree/master/notebooks) to help users integrate the raspberry pi and `pirecorder` in their work.
 
 ![Screenshots of pirecorder in action, from configuring the camera with the interactive video stream, running recordings, testing and scheduling future recordings, and converting recorded media.](Figure1.jpg)
 
[…]
