### 2020-09-16 #2584 @lucask07
> > > 1. In the convert function I might have discovered a tiny bug.
> > >    from pirecorder import Convert
> > >    Convert(indir = "/home/pi/Videos", outdir = "/home/pi/Videos/converted", resizeval = 0.5)
> > >    Convert(indir = "/home/pi/Videos", outdir = "/home/pi/Videos/converted", resizeval = 0.5, withframe = True)
> > >    when i do the "withframe = True" then the result becomes a scrambled video mostly grey and random pix, but it works without.
> >
> >
> > @DerJH I also cannot reproduce this issue. When the `withframe` parameter is set to `True` then instead of fast ffmpeg conversion, opencv will be used for conversion (and drawing the frame number on each frame of the video). So I think there is an issue with the version of opencv you are using. Python2.7 is no longer supported and I think it could well be that your version of opencv is not up to par anymore. Again I recommend to try again on python3 and to use a virtual environment to install the right version of opencv. I wrote a special documentation page for that, but it should hopefully work after running this simple command: `pip3 install opencv-contrib-python==4.1.0.25`.
>
> I am also convinced that this is some installation hiccup with python and opencv on my side. I will try to check this on a fresh install in the coming days. But given that it worked for @lucask07 it should not hinder the publication.

Hi @DerJH, unfortunately, I do not have a Raspberry Pi and so am not able to test for functionality.
