### 2020-09-16 #2584 @lucask07
@JolleJolles
Looks great that most issues are closed. I'm OK with signing off on publication once the wording on **camera calibration** is cleared up in the paper.  Note that I will need to leave the functionality verification box unchecked since I lack the hardware.
