### 2020-09-17 #2584 @JolleJolles
> Hi @JolleJolles,
> I'm going through the paper and have a few concerns/nitpicks that I'd suggest you clarify.
>
> * The paper references "get the optimal camera shutter speed". This is a rather complex calculation that attempts to limit saturation while simultaneously having the darkest pixels collect the most possible light. I can't find any auto exposure calculations in your code. Can you clarify how the optimal shutter speed is calculated/retrieved?
In this context, it was meant optimal in terms of the user's perspective. I rewrote the sentence to say "...set the ideal camera shutter speed" to make this clearer.

> * The paper reference “Calibrating the camera.” This is not appropriate usage of the term **_calibrate_** for a camera. Camera calibration may involve determining the focal length and image center or estimating the lens distortion. Other calibration efforts may center around aspects of sensor performance such as measuring the color correction matrix or locating hot pixels. This does not appear to be what _pirecorder_ is doing. I think your package enables real-time streaming so that a user is given feedback while positioning a camera? Can you change the paper wording to clarify what _pirecorder_ helps the user accomplish?
One of the definitions of "calibrate" is to "carefully assess, set, or adjust". This is what is meant here. By means of an interactive user stream, both with the `stream` and `camconfig` functions, the user is for example able to carefully position the camera and center it, adjust the camera focus using a zoomed-in view of the camera, and adjust external lighting and camera sensitivity settings to get the ideal image brightness and shutter speed. I, therefore, do feel that the term "calibrate" is correctly used here. But I agree that this was not clear enough in the paper. I have therefore carefully revised the paper to better clarify what pirecorder helps accomplish in terms of the steps until recording.

> * _Paper grammar_: “Multiple configuration files can be used that can be edited directly, such as to set the video duration, quality, or number of images, or dynamically via an interactive video stream, such as to draw the region of interest, set the desired white-balance, or get the optimal camera shutter speed.”
>   The “or”s in this list are confusing. Should these actually be “and”s? Consider splitting this sentence into 2 or 3 sentences for clarity.
I have revised this paragraph to improve grammar and readability, including by splitting the sentence into multiple sentences.

> * The [Jupyter notebook](https://github.com/JolleJolles/pirecorder/blob/9b54d9cd77996e5387072d5be3313f804782d564/notebooks/running-PiRecorder.ipynb) references the three possible cameras to use:
>
> 1. v1.5 camera model (2592x1944)
> 2. v2 camera (3280x2464)
> 3. HQ camera (4056x3040)
>    I think these 3 cameras should be more clearly called out in the README.
These are the three official cameras, all with their own maximum resolution. But besides these three official cameras there are many unofficial cameras that also connect to the MIPI port and should generally be supported. I have revised the sentences related to this point in the Jupiter notebook to clarify to the user that the maximum resolution depends on the camera used and that "for example" the three official cameras have max resolutions of (2592,1944), (3280x2464) and (4056x3040). I have also renamed the `cameratype` parameter to `maxres` with the ability to set the resolution via a tuple as well as a string detailing either "v1.5", "v2", or "hq", to help users get the maximum out of their camera.

> * The code _camconfig.py_ allows for a limited range of ISO.
>
> ```
> isos = [100,200,320,400,500,640,800]
> ```
>
> Is this consistent with the capabilities of the 3 camera modules? Is it possible to probe the camera module for the available ISO settings?
See my reply in a previous comment.

> * The code _camconfig.py_ appears to limit resolution to 1640x1232:
>
> ```
> if (res[0]*res[1])>=(1640*1232):
>         res = maxrect(dims, maxdims = (1640,1232), decimals = 0)
> cam.resolution = picamconv(res)
> ```
>
> Is my interpretation correct? Is this limitation required?
Yes this limitation was included to prevent users from using a too high resolution for streaming and getting a shuttering video. The limitation could be removed but most users want to use a smaller video with good frame rate anyway to calibrate the camera settings as that is what the `camconfig` function is about.
