### 2020-09-17 #2584 @JolleJolles
> @JolleJolles
> Looks great that most issues are closed. I'm OK with signing off on publication once the wording on **camera calibration** is cleared up in the paper. Note that I will need to leave the functionality verification box unchecked since I lack the hardware.

I have revised the paper in a number of ways, including removing the word "calibration" and adding a couple of sentences that clarify the streaming functionalities among some other changes.
