### 2020-09-28 #2584 @marcosvital
Good, all references are fine.

@JolleJolles, I'll generate another manuscript proof, so you can take a carefull (maybe final?) look. I'll read it myself too, and will let you know if I find any issues, ok?
