### 2020-09-28 #2584 @JolleJolles
> Good, all references are fine.
>
> @JolleJolles, I'll generate another manuscript proof, so you can take a carefull (maybe final?) look. I'll read it myself too, and will let you know if I find any issues, ok?

Hi @marcosvital, I have read the proofs and they look good. I just have three small changes, all for the last paragraph:
- one small spelling mistake to the very last sentence ("their" should be "the")
- for the references in the last paragraph, fix order (2018 should be before 2019) and change reference jolles et al 2017 by 2020 (rather than add both of them as I don't want to over cite my own studies)
- add "(in prep") for the description of the study "more recently, the autonomous long-term recording of fish each day, every day, for the first four-month of their life"

I will wait to hear if you have any final corrections to make and then commit these changes. I will then also generate a final release of the paper and publish that to pypi.
