### 2020-09-29 #2584 @marcosvital
@JolleJolles, everything seems fine, and I would like to ask for one more change: please include the "statement of need" as a titled paragraph. I believe that could be done with your second paragraph, as it seems to cover this content. It should look [like in this paper](https://joss.theoj.org/papers/10.21105/joss.02607).
