+++ paper/paper.bib
@@ -73,6 +73,17 @@
     volume = {154},
     year = {2019}
 }
+@article{Jolles2020,
+    author = {Jolles, Jolle W and Mazu{\'{e}}, Geoffrey P F and Davidson, Jacob and Behrmann-Godel, Jasminca and Couzin, Iain D},
+    doi = {10.1038/s41598-020-69057-0},
+    journal = {Scientific Reports},
+    pages = {12282},
+    title = {{{\textless}i{\textgreater}Schistocephalus{\textless}/i{\textgreater} parasite infection alters sticklebacks' movement ability and thereby shapes social interactions}},
+    volume = {10},
+    number = {1},
+    year = {2020}
+}
+
 @misc{Jones2017,
     author = {Jones, D},
     title = {{Picamera 1.13 Documentation}},
