+++ setup.py
@@ -27,7 +27,7 @@ DISTNAME="pirecorder"
 MAINTAINER="Jolle Jolles"
 MAINTAINER_EMAIL="j.w.jolles@gmail.com"
 URL="https://github.com/jollejolles"
-DOWNLOAD_URL="https://github.com/jollejolles/pirecorder/archive/v3.1.0.tar.gz"
+DOWNLOAD_URL="https://github.com/jollejolles/pirecorder/archive/v3.2.0.tar.gz"
 
 with open("README.md") as f:
     readme = f.read()
