### 2020-09-29 #2584 @JolleJolles
I have incorporated your requested change and committed my suggested changes to the paper.
