+++ paper/paper.md
@@ -31,12 +31,12 @@ So far, researchers have often relied on writing their own recordings scripts to
 
 ![Screenshots of pirecorder in action, from configuring the camera with the interactive video stream, running recordings, testing and scheduling future recordings, and converting recorded media.](Figure1.jpg)
 
-A core functionality of `pirecorder` is that it works with configuration files. These files make it possible to store a wide range of camera and recording settings that are then automatically used for recordings without requiring further user input. Furthermore, multiple configuration files can be stored and used, such as to easily start recordings for different experimental contexts or treatments. Configuration files can be edited directly, or parameters can be set in python or using the interactive video stream functionalities. Recordings can be easily initiated remotely, such as via an SSH connection, and scheduled to automatically start and stop at specific times in the future. By its use of configuration files and the automatic naming of files, `pirecorder` also makes it possible to start controlled recordings on multiple raspberry pi's simultaneously, such as with [csshX](https://github.com/brockgr/csshx), which sends the same command to multiple computers at once.  
+A core functionality of `pirecorder` is that it works with configuration files. These files make it possible to store a wide range of camera and recording settings that are then automatically used for recordings without requiring further user input. Furthermore, multiple configuration files can be stored and used, such as to easily start recordings for different experimental contexts or treatments. Configuration files can be edited directly, or parameters can be set in python or using the interactive video stream functionalities. Recordings can be easily initiated remotely, such as via an SSH connection, and scheduled to automatically start and stop at specific times in the future. By its use of configuration files and the automatic naming of files, `pirecorder` also makes it possible to start controlled recordings on multiple raspberry pi's simultaneously, such as with [csshX](https://github.com/brockgr/csshx), which sends a command to multiple computers at once.  
 
 # Use cases
-`pirecorder` has already been used successfully in a number of studies, such as to facilitate the high-throughput recording of large numbers of individuals and shoals of fish [@Jolles2018; @Jolles2019, @Jolles2020] and more recently, the autonomous long-term recording of fish each day, every day, for the first four-month of their life (in prep). By facilitating and streamlining controlled and automated image and video recordings, I hope `pirecorder` will increasingly help scientists simplify and improve the collection of high quality data and thereby ultimately enhance their research.
+`pirecorder` has already been used successfully in a number of studies, such as to facilitate the high-throughput recording of large numbers of individuals and shoals of fish [@Jolles2018; @Jolles2019; @Jolles2020] and more recently, the autonomous long-term recording of fish each day, every day, for the first four-month of their life (in prep). By facilitating and streamlining controlled and automated image and video recordings, I hope `pirecorder` will be used by scientists to help simplify and improve their collection of high quality data and thereby ultimately enhance their research.
 
 # Acknowledgements
-This work was made possible by an Alexander von Humboldt Postdoctoral Fellowship, a Zukunftskolleg Postdoctoral Fellowship, and a research grant from the Dr. J.L. Dobberke Foundation. I would also like to thank Lucas Koerner and Jan Heuschele for helpful feedback on the paper and the package.
+I would like to thank Lucas Koerner and Jan Heuschele for helpful feedback on the package and paper. This work was made possible by a Postdoctoral Fellowship from the Alexander von Humboldt Foundation, a  Postdoctoral Fellowship from the Zukunftskolleg, Institute of Advanced Study at the University of Konstanz, and a research grant from the Dr. J.L. Dobberke Foundation.
 
 # References
