+++ paper/paper.bib
@@ -78,7 +78,7 @@
     doi = {10.1038/s41598-020-69057-0},
     journal = {Scientific Reports},
     pages = {12282},
-    title = {{{\textless}i{\textgreater}Schistocephalus{\textless}/i{\textgreater} parasite infection alters sticklebacks' movement ability and thereby shapes social interactions}},
+    title = {{Schistocephalus parasite infection alters sticklebacks' movement ability and thereby shapes social interactions}},
     volume = {10},
     number = {1},
     year = {2020}
