### 2020-09-29 #2584 @JolleJolles
I noticed some small errors with the references so fixed them now.
