### 2020-09-29 #2584 @JolleJolles
Okay the final proofs looks good. I have committed the final changes and release pirecorder 3.2.0, which is now also available on pypi.
