### 2020-10-06 #2584 @JolleJolles
@marcosvital v3.2.0 is indeed the last release and it is available on Zenodo (https://zenodo.org/record/4058628#.X3y8vZMzZoA)
Thanks
