### 2020-10-07 #2584 @kthyng
Hi @JolleJolles! I will help you through the rest of the publishing process. First, please edit the metadata for your Zenodo archive so that the title and author list exactly match your JOSS paper.
