### 2020-10-08 #2584 @JolleJolles
> Hi @JolleJolles! I will help you through the rest of the publishing process. First, please edit the metadata for your Zenodo archive so that the title and author list exactly match your JOSS paper.

Hi @kthyng, I have updated the Zenodo archive. I hope it is okay now? Otherwise let me know what else I need to change. Thanks
