### 2020-10-09 #2584 @JolleJolles
> Congrats to @JolleJolles on your new publication! Thanks to editor @marcosvital and reviewers @lucask07 and @DerJH for making all of this possible.
>
> (I will close this issue once the doi resolves)

Wonderful! Thanks all for your help with this, I am very excited my paper is now published in JOSS.
