### 2020-10-10 8573cfc
Video and JOSS paper added
