### 2020-10-10 f4e40af
Video and JOSS paper added improvements
